leds: aw2013_led: enable regulators to fix failure when light led
The suspend routine releases all the regulators acquired by the
driver if no LED is enabled. However, this causes an issue when
later the user-space tries to enable the LED. Add a check for
this effect in blinking routine to check if regulators are
disabled and then enable them in case they are.

This patch is propagated from msm-3.10 kernel.
(commit: Id0ee80f37bc779cb3cbe4f1291ece92d59efdf6f)
leds: aw2013_led: enable regulators to fix failure when light led

Change-Id: Id0ee80f37bc779cb3cbe4f1291ece92d59efdf6f
Signed-off-by: Mao Li <maol@codeaurora.org>
Signed-off-by: Shantanu Jain <shjain@codeaurora.org>
Signed-off-by: SunnyRaj84348 <sunnyraj7461@gmail.com>
Signed-off-by: dndxtz <dindascape@gmail.com>