leds: aw2013_led: disable regulators when not needed
Turn off the LED regulators if brightness of all the LEDs that are
connected to the host is 0. This change is required to save the power
consumption.

This patch is propagated from msm-3.10 kernel.
(commit: I9bae343becb9a794a5ecc8ddf9c2edb3b5aecfce)
leds: aw2013_led: disable regulators when not needed

Change-Id: I9bae343becb9a794a5ecc8ddf9c2edb3b5aecfce
Signed-off-by: Mao Li <maol@codeaurora.org>
Signed-off-by: Shantanu Jain <shjain@codeaurora.org>
Signed-off-by: SunnyRaj84348 <sunnyraj7461@gmail.com>
Signed-off-by: dndxtz <dindascape@gmail.com>